requirements.txt
flask==3.0.3
selenium==4.23.1
gunicorn==21.2.0
requests==2.32.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
